+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Changed
 
 - Upgrade Marp Core to [v0.8.0](https://github.com/marp-team/marp-core/releases/v0.8.0) ([#23](https://github.com/marp-team/marp-vscode/pull/23))
+- Upgrade dependent packages to the latest version ([#24](https://github.com/marp-team/marp-vscode/pull/24))
 
 ## v0.1.0 - 2019-03-22